+++ src/frame_info.rs
@@ -15,8 +15,8 @@ fn fletcher16(data: &[u8]) -> u16 {
     let mut sum1: u16 = 0;
     let mut sum2: u16 = 0;
 
-    for index in 0..data.len() {
-        sum1 = (sum1 + data[index] as u16) % 255;
+    for byte in data {
+        sum1 = (sum1 + *byte as u16) % 255;
         sum2 = (sum2 + sum1) % 255;
     }